Move hooks to render method
