+++ django_reactive/forms/widgets.py
@@ -1,4 +1,5 @@
 import json
+import logging
 
 from django.forms import Media
 from django.forms.widgets import Widget
@@ -6,13 +7,17 @@ from django.template.loader import render_to_string
 from django.utils.safestring import mark_safe
 
 
+logger = logging.getLogger(__name__)
+
+
 class ReactJSONSchemaFormWidget(Widget):
 
     template_name = "django_reactive.html"
 
-    def __init__(self, schema, ui_schema=None, extra_css=None, extra_js=None, **kwargs):
+    def __init__(self, schema, ui_schema=None, hooks=None, extra_css=None, extra_js=None, **kwargs):
         self.schema = schema
         self.ui_schema = ui_schema
+        self.hooks = hooks
         self.extra_css = extra_css
         self.extra_js = extra_js
         super(ReactJSONSchemaFormWidget, self).__init__(**kwargs)
@@ -34,6 +39,11 @@ class ReactJSONSchemaFormWidget(Widget):
         return Media(css={"all": css}, js=js)
 
     def render(self, name, value, attrs=None, renderer=None):
+        if self.hooks:
+            try:
+                [_(self.schema, self.ui_schema) for _ in self.hooks]
+            except BaseException as exc:
+                logger.error("Error applying JSON schema hooks: %s", exc, exc_info=True)
         context = {
             "data": value,
             "name": name,
