+++ example/tester/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.contrib.admin import ModelAdmin
 
-from .models import TestModel, Todo
+from .models import TestModel
 from .forms import TestModelForm
 
 
@@ -10,10 +10,3 @@ class TestModelAdmin(ModelAdmin):
 
 
 admin.site.register(TestModel, TestModelAdmin)
-
-
-class TodoAdmin(ModelAdmin):
-    pass
-
-
-admin.site.register(Todo, TodoAdmin)
