+++ tests/testapp/models.py
@@ -3,18 +3,18 @@ from django.db import models
 from django_reactive.fields import ReactJSONSchemaField
 
 
-
 def modify_max_length(schema, ui_schema):
     import random
+
     max_length = random.randint(20, 30)
-    schema['properties']['test_field']['maxLength'] = max_length
-    ui_schema['test_field']['ui:help'] = f"Field maximum: {max_length} characters"
+    schema["properties"]["test_field"]["maxLength"] = max_length
+    ui_schema["test_field"]["ui:help"] = f"Field maximum: {max_length} characters"
 
 
 def maybe_disable_field(schema, ui_schema):
     some_condition = 1
     if some_condition:
-        ui_schema['another_test_field'] = {"ui:disabled": True}
+        ui_schema["another_test_field"] = {"ui:disabled": True}
 
 
 class HookedSchemaModel(models.Model):
@@ -38,7 +38,8 @@ class HookedSchemaModel(models.Model):
         ui_schema={
             "test_field": {"ui:help": "Field maximum: 10 characters"},
         },
-        hooks=[modify_max_length, maybe_disable_field])
+        hooks=[modify_max_length, maybe_disable_field],
+    )
 
 
 class SchemaModel(models.Model):
